Allow to specify feature without value in condition for toolset.flags call.
Patch from Alexey Pakhunov.

What this patch does is allows to match absense of optional feature in
the 'flags' rule. For example:

  flags msvc .SETUP <architecture>/<address-model>64   : " x86_amd64" ;
  flags msvc .SETUP <architecture>ia64/<address-model> : " x86_ia64" ;

This will produce "x86_amd64" when the <address-model> is set to 64
and <architecture> is not set. Likewise, for <architecture>ia64 and unset
address model.

Without this patch, we'd have two choices:

  - adding 'default' to the list of feature values. But that would
    add 'architecture-default' to the target path. Ick!

  - creating top-level variable .SETUP with default value, and only
    matching configurations with fully-specified feature values. But
    this won't handle the case above: we really need to check
    which one of two features is unspecified.


[SVN r31036]